put geofield and quantity models back into log.dart
